+++ ruma-client-api/CHANGELOG.md
@@ -19,6 +19,7 @@ Breaking changes:
 * Fix event types in `r0::context::get_context`
 * Fix event types in `r0::sync::sync_events`
 * Update type of `user_id` in `r0::account::whoami` from `String` to `ruma_identifiers::UserId`
+* Update type of `limited` in `r0::sync::sync_events::Timeline` from `Option<bool>` to `bool`
 
 Improvements: